Cleaup: Remove isSettingEnabled setting util
This removes the unnessary util that I created in a previous commit.
I updated the isModuleEnabled method that is provided by useSettings to provide a function
to allow the results of any settings updates to be recalculated.
Cleans up a lot of unnessary effects and signals that weren't needed in
the first place.